+++ generator/parser/src/parsing/payload/attr/mod.rs
@@ -35,8 +35,6 @@ impl Parse for PayloadAttrs {
                         let as_str = ident.to_string();
                         if as_str == PayloadAttrId::Include.to_string() {
                             attrs.push(PayloadAttr::Include)
-                        } else if as_str == PayloadAttrId::Ctx.to_string() {
-                            attrs.push(PayloadAttr::Ctx)
                         } else if as_str == PayloadAttrId::NoDefaultSig.to_string() {
                             attrs.push(PayloadAttr::NoDefaultSig)
                         } else if as_str == PayloadAttrId::Bincode.to_string() {