+++ generator/parser/src/tys/mod.rs
@@ -1,5 +1,6 @@
 mod block;
 mod config;
+mod context;
 mod derives;
 mod modpath;
 mod payload;
@@ -7,6 +8,7 @@ mod vis;
 
 pub use block::*;
 pub use config::*;
+pub use context::*;
 pub use derives::*;
 pub use modpath::*;
 pub use payload::*;