+++ generator/parser/src/tys/payload/attr/mod.rs
@@ -59,9 +59,6 @@ impl PayloadAttrs {
     pub fn is_crypt(&self) -> bool {
         self.0.iter().any(|attr| matches!(attr, PayloadAttr::Crypt))
     }
-    pub fn is_ctx(&self) -> bool {
-        self.0.iter().any(|attr| matches!(attr, PayloadAttr::Ctx))
-    }
     pub fn is_include(&self) -> bool {
         self.0
             .iter()
@@ -73,7 +70,6 @@ impl PayloadAttrs {
 pub enum PayloadAttr {
     Path(ModulePath),
     Include,
-    Ctx,
     NoDefaultSig,
     Hooks,
     NoAutoCrc,
@@ -90,7 +86,6 @@ impl fmt::Display for PayloadAttr {
             match self {
                 Self::Path(path) => format!("{}({path})", self.id()),
                 Self::Include => PayloadAttrId::Include.to_string(),
-                Self::Ctx => PayloadAttrId::Ctx.to_string(),
                 Self::NoDefaultSig => PayloadAttrId::NoDefaultSig.to_string(),
                 Self::Hooks => PayloadAttrId::Hooks.to_string(),
                 Self::NoAutoCrc => PayloadAttrId::NoAutoCrc.to_string(),